Introduce TermServer and make the telnet impl
