+++ src/main/java/io/vertx/ext/shell/term/SSHOptions.java
@@ -30,7 +30,7 @@
  *
  */
 
-package io.vertx.ext.shell.net;
+package io.vertx.ext.shell.term;
 
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.core.json.JsonObject;
