+++ src/test/java/io/vertx/ext/shell/term/VertxTelnetHandlerTest.java
@@ -30,32 +30,24 @@
  *
  */
 
-package io.vertx.ext.shell.net;
+package io.vertx.ext.shell.term;
 
 import io.termd.core.telnet.TelnetHandler;
-import io.termd.core.tty.TelnetTtyTestBase;
+import io.termd.core.telnet.TelnetHandlerTest;
 
 import java.io.Closeable;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
 /**
+ * See <a href="http://commons.apache.org/proper/commons-net/examples/telnet/TelnetClientExample.java>for more possibilities</a>
+ *
  * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
  */
-public class VertxBinaryTelnetTtyTest extends TelnetTtyTestBase {
-
-  public VertxBinaryTelnetTtyTest() {
-    binary = true;
-  }
+public class VertxTelnetHandlerTest extends TelnetHandlerTest {
 
   @Override
   protected Function<Supplier<TelnetHandler>, Closeable> serverFactory() {
     return VertxTelnetTermTest.VERTX_SERVER;
-  }
-
-  @Override
-  protected void assertThreading(Thread connThread, Thread schedulerThread) throws Exception {
-    assertTrue(connThread.getName().startsWith("vert.x-eventloop-thread"));
-    assertEquals(connThread, schedulerThread);
   }
 }
